Fix: 수정사항 추가
 healthbar 위치,  hackinggoal 등

+++ EPOH/Assets/ScriptableObject/BossData/BossData.cs
@@ -8,7 +8,7 @@ public class BossData : ScriptableObject
     [Header("보스 정보")]
     public string name;
     public int bossIndex;
-    public float hackingGoal;
+    public int hackingGoal;
     public string clearEventId;
     public string failEventId;
     [SerializeField] private float _maxHealth;
